11.10
-fixed whip attack animation playing but not doing dmg:
the problem is that whip bomb is set to corrode damage (which is only
armor damage for marines).
The fix is to add in a filter that prevents the animation of whip bomb
from playing, if the target
is a marine with 0 armor. Because why play the animation if no dmg
occurs? (One could argue
for the splash damage?)

i also adjusted whips and hydras to not attack open doors that are
already broken

adjusted arcs to not fire before doors open (correctly now)